refactor(sever) : refactor format api category and add GlobalExceptionHandler

+++ src/main/java/com/example/OrderCoffeeBE/Service/impl/CategoryServiceImpl.java
@@ -2,12 +2,14 @@ package com.example.OrderCoffeeBE.Service.impl;
 
 import com.example.OrderCoffeeBE.Entity.Category;
 import com.example.OrderCoffeeBE.Service.CategoryService;
+import com.example.OrderCoffeeBE.Util.Error.ResourceNotFoundException;
 import com.example.OrderCoffeeBE.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+
+import static org.springframework.data.jpa.domain.AbstractPersistable_.id;
 
 @RequiredArgsConstructor
 @Service("categoryService")
@@ -21,12 +23,23 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category findByIdCate(int id) {
-        Optional<Category> categoryOptional = categoryRepository.findById(id);
-        return categoryOptional.orElse(null);
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Category not found with id " + id));
     }
 
     @Override
     public Category createCate(Category Category) {
+        if(Category == null)
+        {
+            throw new ResourceNotFoundException("Category Is Required");
+        }
+        var isNameExist = categoryRepository.findByName(Category.getName());
+        if (isNameExist != null) {
+            throw new ResourceNotFoundException("Category " + Category.getName() + " already exists, please use another category");
+        }
+        if(Category.getName() == null || Category.getName().isEmpty()) {
+            throw new ResourceNotFoundException("User is not empty");
+        }
         return categoryRepository.save(Category);
     }
 
@@ -34,19 +47,24 @@ public class CategoryServiceImpl implements CategoryService {
     public Category updateCate(Category updateCategory) {
         Category currentCate = this.findByIdCate(updateCategory.getId());
         if (currentCate == null) {
-            return null;
+             throw new ResourceNotFoundException("Category is required");
         }
+        var isNameExist = categoryRepository.findByName(updateCategory.getName());
+        if (isNameExist != null) {
+            throw new ResourceNotFoundException("Category " + updateCategory.getName() + " already exists, please use another category");
+        }
+        //update cate
         currentCate.setName(updateCategory.getName());
+        //save cate
         return this.categoryRepository.save(currentCate);
     }
 
     @Override
     public void deleteCate(int id) {
+        var category = categoryRepository.findById(id).orElse(null);
+        if (category == null) {
+            throw new ResourceNotFoundException("Category not found");
+        }
         this.categoryRepository.deleteById(id);
-    }
-
-    @Override
-    public boolean isNameExist(String name) {
-        return this.categoryRepository.existsByName(name);
     }
 }
